Added check if user is logged in or not

+++ app/Http/Controllers/LoginController.php
@@ -72,7 +72,9 @@ class LoginController extends Controller
     public function destroy(Request $request)
     {
         // Remove Shopping Session from Database
-        $this->shoppingSessionService->deleteByUserId(Auth::user()->id);
+        if (Auth::check()) {
+            $this->shoppingSessionService->deleteByUserId(Auth::user()->id);
+        }
         Auth::guard('web')->logout();
         $request->session()->invalidate();
         $request->session()->regenerateToken();
